% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{damnfine}
\alias{damnfine}
\title{A Twin Peaks palette generator}
\usage{
damnfine(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{RedRoom}, \code{LumberMill},  \code{PhoneBooth},
\code{PeteMartell}, \code{OneEyedJacks}}

\item{n}{Number of colors desired. RedRoom1, LumberMill, PhoneBooth,
PeteMartell and OneEyedJacks color schemes are from the video
 \href{https://vimeo.com/218231355}{Color by Numbers: Twin Peaks}.. BeckyBriggs and
RedRoom2 are from the instagram
\href{https://www.instagram.com/colorpalette.cinema/}{colorpalette.cinema}
  If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are color palettes from the television show Twin Peaks and the movie
Fire Walk with Me.
}
\examples{
damnfine("RedRoom")
damnfine("OneEyedJacks")
damnfine("LumberMill")
damnfine("LumberMill", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- damnfine(21, name = "PhoneBooth", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
