% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R, R/damnfinecolors.R
\docType{data}
\name{damnfinecolors}
\alias{damnfinecolors}
\alias{damnfinecolors-package}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 12.}
\usage{
damnfinecolors
}
\description{
Use \code{\link{damnfine}} to construct palettes of desired length.

list of palettes from Twin Peaks
}
\details{
list of palettes from Twin Peaks
}
\keyword{datasets}
